🔥 Add Housing Model

+++ train_iris.py
@@ -2,6 +2,7 @@
 import pickle
 from sklearn.datasets import load_iris
 from sklearn.ensemble import RandomForestClassifier
+import os
 
 iris = load_iris()
 X, y = iris.data, iris.target
@@ -9,5 +10,9 @@ X, y = iris.data, iris.target
 model = RandomForestClassifier()
 model.fit(X, y)
 
-with open("app/model.pkl", "wb") as f:
-    pickle.dump(model, f)
+
+filename = "app/model_iris.pkl"
+os.makedirs(os.path.dirname(filename), exist_ok=True)
+if not os.path.isfile(filename):
+    with open(filename, "wb") as f:
+        pickle.dump(model, f)
